Refactor based on PR
